Create K’th Non-repeating Character in Python.py
K’th Non-repeating Character in Python using List Comprehension and OrderedDict.
Given a string and a number k, find the k-th non-repeating character in the string. Consider a large input string with lacs of characters and a small character set. How to find the character by only doing only one traversal of input string?
My code is the solution to this question